LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-transition-inherited-used-by-standard-property-expected.txt


FAIL Running a transition an inherited CSS variable is reflected on a standard property using that variable as a value assert_equals: expected "150px" but got "100px"
